Refactor video streaming components and update dependencies
Refactor video player components for GCP integration, update React types
to latest versions, and add Vidstack React Tailwind plugin. This commit
also includes updated package-lock.json and tailwind.config.ts files.
