+++ app/globals.css
@@ -8,6 +8,8 @@ body,
   height: 100%;
   padding: 0;
   margin: 0;
+  --media-brand: 245 245 245;
+  --media-focus: 78 156 246;
 }
 
 @layer base {
