+++ components/courses/courseId/chapters/chapterId/VideoPlayer.tsx
@@ -9,9 +9,10 @@ import { Loader2, Lock } from "lucide-react";
 
 import { cn } from "@/lib/utils";
 import { useConfettiStore } from "@/hooks/useConfettiStore";
+import { Player } from "@/components/video/VideoPlayer";
 
 interface VideoPlayerProps {
-  playbackId: string;
+  videoUrl: string;
   courseId: string;
   chapterId: string;
   nextChapterId?: string;
@@ -21,7 +22,7 @@ interface VideoPlayerProps {
 }
 
 export const VideoPlayer = ({
-  playbackId,
+  videoUrl,
   courseId,
   chapterId,
   nextChapterId,
@@ -40,7 +41,7 @@ export const VideoPlayer = ({
           `/api/courses/${courseId}/chapters/${chapterId}/progress`,
           {
             isCompleted: true,
-          }
+          },
         );
 
         if (!nextChapterId) {
@@ -67,20 +68,21 @@ export const VideoPlayer = ({
         </div>
       )}
       {isLocked && (
-        <div className="absolute inset-0 flex items-center justify-center bg-slate-800 flex-col gap-y-2 text-secondary">
+        <div className="absolute inset-0 flex flex-col items-center justify-center gap-y-2 bg-slate-800 text-secondary">
           <Lock className="h-8 w-8" />
           <p className="text-sm">This chapter is locked</p>
         </div>
       )}
       {!isLocked && (
-        <MuxPlayer
-          title={title}
-          className={cn(!isReady && "hidden")}
-          onCanPlay={() => setIsReady(true)}
-          onEnded={onEnd}
-          autoPlay
-          playbackId={playbackId}
-        />
+        // <MuxPlayer
+        //   title={title}
+        //   className={cn(!isReady && "hidden")}
+        //   onCanPlay={() => setIsReady(true)}
+        //   onEnded={onEnd}
+        //   autoPlay
+        //   videoUrl={videoUrl}
+        // />
+        <Player title={title} url={videoUrl} onEnded={onEnd} />
       )}
     </div>
   );
