+++ components/video/layout/VideoLayout.tsx
